# docs/rec_candidate_register_api.openapi.yaml
openapi: 3.0.3
info:
  title: Digify ERP — Career Portal Candidate API
  version: 2.0.0
  description: |
    Token-free career portal endpoints under `/api/candidate` (no JWT).

    - **Register:** `REC.CANDIDATE_USER_PKG.REGISTER_CANDIDATE_USER`
    - **Login:** `REC.CANDIDATE_LOGIN_PKG.LOGIN_CANDIDATE` — validates credentials (no JWT issued)

    Related public career APIs (ERP staff JWT not required):
    - `GET /api/rec/job-postings` — list postings
    - `GET /api/rec/job-postings/{postingGuid}` — posting detail
    - `POST /api/rec/job-postings/{postingGuid}/apply` — submit application (multipart + resume_file)
    - `GET /api/candidate/offers` — list EXTENDED job offers for a candidate
    - `GET /api/candidate/offers/{offerGuid}` — EXTENDED offer detail
    - `POST /api/rec/job-offers/{offerGuid}/accept` — accept offer
    - `POST /api/rec/job-offers/{offerGuid}/decline` — decline offer

    Passwords on register are hashed server-side (Argon2id) before Oracle storage.

    ## Hostname tenant resolution (preferred)

    Prefer tenant hostnames such as `{tenant}.careers.digifyhr.com`.
    The API resolves `enterprise_id` via `ENT.ENT_ENTERPRISES_PKG` action `RESOLVE_SUBDOMAIN`.
    Prefer public aliases under `/api/public/*` (no `enterprise_id` query/body).
    Client-supplied `enterprise_id` is deprecated when a tenant hostname is present.
    See `docs/hostname_enterprise_resolution.md`.
servers:
  - url: /
    description: Application base URL
tags:
  - name: Career Portal
    description: Public candidate registration
paths:
  /api/candidate/register:
    post:
      tags: [Career Portal]
      summary: Register candidate user (career portal)
      operationId: registerCandidateUser
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CandidateRegisterMultipart'
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateRegisterJson'
      responses:
        '200':
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterSuccess'
              example:
                success: true
                message: Candidate user registered successfully.
                candidate_id: 101
                candidate_guid: 5332C3095B5F68E5E0638711000A5963
                candidate_user_id: 10
                candidate_user_guid: 5332C3095B5F68E5E0638711000A5964
        '400':
          description: Validation or package error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
              example:
                success: false
                message: Resume upload failed.
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
  /api/candidate/login:
    post:
      tags: [Career Portal]
      summary: Login candidate user (career portal)
      operationId: loginCandidateUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateLoginRequest'
            example:
              enterprise_id: 1
              email: johnny.doe@example.com
              password: Password1!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateLoginSuccess'
        '401':
          description: Invalid credentials or inactive user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
              examples:
                invalidCredentials:
                  value:
                    success: false
                    message: Invalid email or password.
                inactiveUser:
                  value:
                    success: false
                    message: Candidate user is not active.
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
  /api/candidate/offers:
    get:
      tags: [Career Portal]
      summary: List EXTENDED job offers for candidate (career portal)
      operationId: listCandidateExtendedOffers
      security: []
      description: |
        Returns job offers with `EXTENDED` status for the given candidate.
        Reads from `REC.V_JOB_OFFER_MANAGEMENT`. No JWT required.
      parameters:
        - $ref: '#/components/parameters/EnterpriseIdQuery'
        - $ref: '#/components/parameters/CandidateGuidQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Offers fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateOfferListResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
  /api/candidate/offers/{offerGuid}:
    get:
      tags: [Career Portal]
      summary: Get EXTENDED job offer detail (career portal)
      operationId: getCandidateExtendedOffer
      security: []
      description: |
        Returns offer detail when the offer is `EXTENDED` and belongs to the candidate.
        No JWT required.
      parameters:
        - $ref: '#/components/parameters/OfferGuidPath'
        - $ref: '#/components/parameters/EnterpriseIdQuery'
        - $ref: '#/components/parameters/CandidateGuidQuery'
      responses:
        '200':
          description: Offer detail fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateOfferDetailResponse'
        '404':
          description: Offer not found or not accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
              example:
                success: false
                message: Offer not found.
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateRegisterError'
components:
  parameters:
    EnterpriseIdQuery:
      name: enterprise_id
      in: query
      required: true
      schema: { type: integer, minimum: 1 }
    CandidateGuidQuery:
      name: candidate_guid
      in: query
      required: true
      schema:
        type: string
        pattern: '^[0-9A-Fa-f]{32}$'
    PageQuery:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitQuery:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
    OfferGuidPath:
      name: offerGuid
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9A-Fa-f]{32}$'
  schemas:
    YnFlag:
      type: string
      enum: [Y, N]
    CandidateRegisterFields:
      type: object
      required: [enterprise_id, first_name, last_name, email, phone, password]
      properties:
        enterprise_id: { type: integer, minimum: 1 }
        first_name: { type: string }
        middle_name: { type: string }
        last_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        password:
          type: string
          minLength: 8
          description: Plain password (hashed server-side; never stored)
        current_title: { type: string }
        current_employer: { type: string }
        years_experience: { type: number }
        current_location: { type: string }
        source: { type: string }
        current_salary: { type: number, minimum: 0 }
        expected_salary: { type: number }
        salary_currency: { type: string }
        notice_period: { type: integer }
        linkedin_profile: { type: string, maxLength: 1000 }
        portfolio_link: { type: string, format: uri, maxLength: 1000 }
        github_link: { type: string, format: uri, maxLength: 1000 }
        willing_to_relocate:
          $ref: '#/components/schemas/YnFlag'
        created_by: { type: string }
    CandidateRegisterMultipart:
      allOf:
        - $ref: '#/components/schemas/CandidateRegisterFields'
        - type: object
          properties:
            resume:
              type: string
              format: binary
              description: Optional resume file
            file:
              type: string
              format: binary
            attachment:
              type: string
              format: binary
            document:
              type: string
              format: binary
    CandidateRegisterJson:
      allOf:
        - $ref: '#/components/schemas/CandidateRegisterFields'
        - type: object
          properties:
            file_name: { type: string }
            file_type: { type: string }
            file_size: { type: integer }
            file_content:
              type: string
              description: Base64-encoded resume (optional)
    CandidateRegisterSuccess:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        candidate_id: { type: integer }
        candidate_guid: { type: string }
        candidate_user_id: { type: integer }
        candidate_user_guid: { type: string }
    CandidateRegisterError:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
    CandidateLoginRequest:
      type: object
      required: [enterprise_id, email, password]
      properties:
        enterprise_id: { type: integer, minimum: 1 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    CandidateLoginSuccess:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: Login successful. }
        data:
          type: object
          properties:
            candidate_user_id: { type: integer }
            candidate_user_guid: { type: string }
            candidate_id: { type: integer }
            candidate_guid: { type: string }
            full_name: { type: string }
            email: { type: string, format: email }
            user_status: { type: string }
    CandidateOfferListItem:
      type: object
      properties:
        offer_guid: { type: string }
        offer_number: { type: string }
        job_title: { type: string }
        location: { type: string }
        work_mode_code: { type: string }
        employment_type_code: { type: string }
        start_date: { type: string, format: date }
        offer_date: { type: string, format: date }
        expiry_date: { type: string, format: date }
        stage: { type: string }
        status_code: { type: string, example: EXTENDED }
        stage_description: { type: string }
        annual_salary: { type: number }
        posting_obj: { type: object }
        position_obj: { type: object }
        department_obj: { type: object }
        grade_obj: { type: object }
        components_json: { type: array, items: { type: object } }
        benefits_json: { type: object }
        terms_json: { type: object }
    CandidateOfferListResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:
          type: array
          items:
            $ref: '#/components/schemas/CandidateOfferListItem'
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
    CandidateOfferDetailResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:
          type: object
          properties:
            offer: { type: object }
            components: { type: array, items: { type: object } }
            benefits: { type: object }
            terms: { type: object }